Add a PDF of the notebook
Even on Github, the notebook does not always render the same way it does
on my system. So, I printed the AdjustedValidate.ipynb notbook as a pdf
file. It is not ideal, since links don't work, and some formatting is
still a little off. But it is readable...something kind of important for
"documentation".