Update FilterAlgorithm
FilterAlgorithm should be usable, even if all the bells and whistles
not there yet. This means:

- at a minimum, allow user to specify arbitrary FIR window
- at a minimum, allow the user to specify decimation, including...
- setting decimation to 1, and expect the same output time stamps
  as the input (this requires proper padding of the input series)
- at a minimum, allow user to change pertinent metadata

This is the first commit to a WIP branch, but it does all the above,
at least via the API. The only item that is not currently possible
via the Controller is the arbitrary FIR window (decimation can be
controlled to a limited extent by tweaking `--interval` and a new
`--filter-interval` options). Doing this would probably require a
config file of some sort to hold filter coefficients, as well as
a few new options. Let's discuss before merging.